fix(protocol): fidelity sweep 2 — ttlMs as schema number; absorb SEP-2577 deprecation markers
(a) ttlMs: the schema declares `number` (@minimum 0); the u64 binding
rejected spec-legal fractional values on deserialize. CacheableResult.ttl_ms
and its embeddings (ListToolsResult, ListResourcesResult,
ListResourceTemplatesResult, ReadResourceResult, ListPromptsResult,
DiscoverResult) are now f64 via a shared serde helper: fractional values
accepted and round-tripped, negative/non-finite rejected, whole values keep
the compact integer wire form (byte-stable for the common ttlMs: 0).
Contract test covers all four behaviors.

(b) SEP-2577 markers: the 2026-06-07 re-pin deprecated the WHOLE Logging
surface — including the per-request _meta logLevel opt-in and the
LoggingLevel enum — and the bindings/rustdoc still claimed those were "the
non-deprecated replacement". Corrected: #[deprecated] now on LoggingLevel
(+ LogLevel alias), META_KEY_LOG_LEVEL, RequestMetaObject.log_level /
with_log_level, ServerCapabilities.logging, ModelHint / ModelPreferences /
ToolChoice, ContentBlock::ToolUse / ToolResult (variants + constructors),
and the sampling/logging trait surface. The behavior remains normative and
implemented through the migration window (earliest removal 2027-07-28);
framework-internal call sites carry scoped #[allow(deprecated)], so the
markers reach downstream consumers without making the framework warn on
itself. Zero warnings on the default build, test build, clippy, and the
2025 opt-in lane builds.

+++ crates/turul-mcp-protocol-2026-07-28/src/discover.rs
@@ -70,7 +70,8 @@ pub struct DiscoverResult {
     pub result_type: ResultType,
 
     /// `CacheableResult.ttlMs` — required by schema (DiscoverResult extends CacheableResult).
-    pub ttl_ms: u64,
+    #[serde(with = "crate::caching::ttl_ms_serde")]
+    pub ttl_ms: f64,
     /// `CacheableResult.cacheScope` — required by schema.
     pub cache_scope: crate::caching::CacheScope,
 
@@ -105,7 +106,7 @@ impl DiscoverResult {
     ) -> Self {
         Self {
             result_type: ResultType::Complete,
-            ttl_ms: 0,
+            ttl_ms: 0.0,
             cache_scope: crate::caching::CacheScope::Public,
             supported_versions,
             capabilities,
@@ -116,7 +117,7 @@ impl DiscoverResult {
     }
 
     /// Set the cache-control hint (`ttlMs` + `cacheScope`).
-    pub fn with_cache(mut self, ttl_ms: u64, cache_scope: crate::caching::CacheScope) -> Self {
+    pub fn with_cache(mut self, ttl_ms: f64, cache_scope: crate::caching::CacheScope) -> Self {
         self.ttl_ms = ttl_ms;
         self.cache_scope = cache_scope;
         self
