+++ crates/turul-mcp-protocol-2026-07-28/src/prompts.rs
@@ -207,7 +207,8 @@ pub struct ListPromptsResult {
     pub next_cursor: Option<Cursor>,
 
     /// `CacheableResult.ttlMs` — required by schema.
-    pub ttl_ms: u64,
+    #[serde(with = "crate::caching::ttl_ms_serde")]
+    pub ttl_ms: f64,
     /// `CacheableResult.cacheScope` — required by schema.
     pub cache_scope: crate::caching::CacheScope,
 
@@ -228,7 +229,7 @@ impl ListPromptsResult {
             result_type: crate::result_type::ResultType::Complete,
             prompts,
             next_cursor: None,
-            ttl_ms: 0,
+            ttl_ms: 0.0,
             cache_scope: crate::caching::CacheScope::Public,
             meta: None,
         }
